+++ src/renderer/i18n/en.ts
@@ -634,9 +634,11 @@ export default {
     personSection: 'Person',
   },
   placeTypes: {
+    continent: 'Continent',
     country: 'Country',
     admin1: 'State/Region',
     province: 'Province',
+    landskap: 'Province',
     county: 'County',
     härad: 'Härad',
     parish: 'Parish',